the implementation of the "Hungarian method" in SciPy works great
to solve the "multiple-keyword-match-problem":

* https://docs.scipy.org/doc/scipy-0.18.1/reference/generated/scipy.optimize.linear_sum_assignment.html
* https://stackoverflow.com/questions/1398822/the-assignment-problem-a-numpy-function
* https://math.stackexchange.com/questions/760855/algorithm-to-compute-maximum-permutation-sum-in-matrix
* https://en.wikipedia.org/wiki/Hungarian_algorithm

However, it is still just a similarity method and we have to reduce
the number of inputs (e.g. remove the elementary flows from the flow
lists to match).

Next step is now to work on the semantic mapping and combining these
both things: searching for terms in a Graph and analyzing their
semantic relations.